+++ Changelog.md
@@ -5,6 +5,12 @@
 
 - Swift 3.0 Support
 - Renamed `HTTPBody` to `httpBody` and `HTTPHeader` to `httpHeader`
+- `HTTPMethod` enum cases are now lowercase
+
+### 1.0.1
+-----------
+
+- Fix `Router` to handle same url with different HTTP methods
 
 ### 1.0.0
 ------------